Refactor all the things!
So I created this project in 2014, it did what I needed it to and I
hadn't looked at it since. However recently I switched to qtile and
gained a use case for dwim's Python API. Around the same time I
received some feedback in the form of issue #1 and pull request #2.

Because of this recent activity, I decided to try and bring this project
up to my current development standards for Python projects. More or less
the only thing missing at this point is a test suite :-).

Amongst the changes I made:

- fixed the bug mentioned in issue #1,
- merged the improvements in pull request #2,
- improved the readme (with an embedded usage message),
- redid the makefile (adding support for wheel distributions),
- improved the Sphinx configuration (changing the theme),
- added code style checks and fixed code style issues,
- bug fix for determine_network_location() (fixed
  exception when all network connections are down)
- improved wait_for_internet_connection()
  with interactive spinner